# Conflicts:
#	app/src/main/java/com/example/myapplication/screens/MainScreen.kt

+++ app/src/main/java/com/example/myapplication/model/LoanResponse.kt
@@ -7,8 +7,8 @@ data class LoanResponse (
 ){
 
     data class Data(
-        val amount: Long,
-        val loanLimit: Long
+        val amount: Int,
+        val loanLimit: Int
     )
 }
